+++ packages/tinacms/src/admin/components/GetCollection.tsx
@@ -136,7 +136,8 @@ const GetCollection = ({
       !allowDelete &&
       // Check there is only one document
       collection.documents?.edges?.length === 1 &&
-      collection.documents?.edges[0]?.node.__typename === 'Document'
+      // Check to make sure the file is not a folder
+      collection.documents?.edges[0]?.node?.__typename !== 'Folder'
     ) {
       const doc = collection.documents.edges[0].node
       handleNavigate(navigate, cms, collection, collectionDefinition, doc)